HttpRequestFailed event sent to mixpanel.

+++ common/Classes/AzoomeeCommon/Analytics/AnalyticsSingleton.cpp
@@ -471,13 +471,13 @@ void AnalyticsSingleton::sessionIdHasChanged(std::string oldSessionId)
     mixPanelSendEvent(eventID, mixPanelProperties);
 }
     
-void AnalyticsSingleton::httpRequestFailed(std::string requestTag, int responseCode, std::string qid)
+void AnalyticsSingleton::httpRequestFailed(std::string requestTag, long responseCode, std::string qid)
 {
     std::string eventID = "httpRequestFailed";
     
     std::map<std::string, std::string> mixPanelProperties;
     mixPanelProperties["requestTag"] = requestTag;
-    mixPanelProperties["responseCode"] = cocos2d::StringUtils::format("%d", responseCode);;
+    mixPanelProperties["responseCode"] = cocos2d::StringUtils::format("%ld", responseCode);;
     mixPanelProperties["qid"] = qid;
     
     mixPanelSendEvent(eventID, mixPanelProperties);
